Fixed main.go
